Dec.4 2025

+++ src/Station_Data/Neighbor.java
@@ -3,11 +3,11 @@ package Station_Data;
 //相邻站点类
 public class Neighbor{
     public Station station;
-    public int distance;
+    public double distance;
 
     public Neighbor prev, next;
 
-    public Neighbor(Station station, int distance){
+    public Neighbor(Station station, double distance){
         this.station = station;
         this.distance = distance;
     }
